refactor: centralize theme and remove hardcoded colors
