+++ react-supabase-auth-template/src/index.js
@@ -1,35 +1,9 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import { ThemeProvider, createTheme } from '@mui/material/styles';
+import { ThemeProvider } from '@mui/material/styles';
+import theme from './theme.js';
 import CssBaseline from '@mui/material/CssBaseline';
 import TabbedAdminDashboard from './components/TabbedAdminDashboard.jsx';
-
-const theme = createTheme({
-  palette: {
-    primary: {
-      main: '#1976d2',
-    },
-    secondary: {
-      main: '#dc004e',
-    },
-    background: {
-      default: '#f5f5f5',
-      paper: '#f5f5f5',
-    },
-  },
-  typography: {
-    fontFamily: '"Roboto", "Helvetica", "Arial", sans-serif',
-  },
-  components: {
-    MuiAppBar: {
-      styleOverrides: {
-        root: {
-          boxShadow: 'none',
-        },
-      },
-    },
-  },
-});
 
 const container = document.getElementById('root');
 if (!container) throw new Error('Failed to find the root element');
