+++ react-supabase-auth-template/src/theme.js
@@ -1,0 +1,30 @@
+import { createTheme } from '@mui/material/styles';
+
+const theme = createTheme({
+  palette: {
+    primary: {
+      main: '#003865',
+    },
+    secondary: {
+      main: '#1565a3',
+    },
+    background: {
+      default: '#f5f5f5',
+      paper: '#f5f5f5',
+    },
+  },
+  typography: {
+    fontFamily: '"Roboto", "Helvetica", "Arial", sans-serif',
+  },
+  components: {
+    MuiAppBar: {
+      styleOverrides: {
+        root: {
+          boxShadow: 'none',
+        },
+      },
+    },
+  },
+});
+
+export default theme;
